Add support Intel SpeedShift.
Intel Speed Shift is Intel's technology to control frequency in
hardware, with hints from software. This renames the hwpstate(4) driver
(for AMD CPUs) to hwpstate_amd(4) and add the hwpstate_intel(4) driver
for Intel CPU that support SpeedShift (Skylake and beyond).

This is Diff 63775 from https://reviews.freebsd.org/D18028.