Untangle APIC virtualization and TPR shadowing.
Bhyves VT-x backend already implements TPR shadowing, but it is bound to
APIC virtualization and APIC virtualization is only available on Xeon
CPUs. This change disconnects TPR shadowing from APIC virtualization and
enables it on all CPUs with support for it.

This accelerates Windows guests running on Intel desktop CPUs by up to
factor 6.