Disable TPR shadowing when the VLAPIC is configured in X2APIC mode.
There might be some corner cases when running in X2APIC mode, that the
code currently doesn't handle. I'm unable to test and debug them,
because the only guest making really use of TPR shadowing is Windows.
Windows required an UEFI ROM to boot and the UEFI ROM doesn't support
X2APIC mode... Therefore play save and disable TPR shadowing as soon
as X2APIC mode is requested.

This is complicated by the fact, that we don't know what VLAPIC mode
will be requested when setting up the VM in vmx_vminit(). Therefore
we're always doing the setup with TPR shadowing enabled and disable
it afterwards.